adversarial_patch_uav_rod/requirements_mmrotate.txt
addict==2.4.0
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
conda==4.9.1
conda-build==3.20.5
conda-package-handling==1.7.0
contourpy==1.1.1
cryptography==2.9.2
cycler==0.12.1
Cython==3.0.12
dataclasses==0.6
decorator==4.4.2
dnspython==2.0.0
e2cnn==0.2.3
filelock==3.0.12
fonttools==4.56.0
future==0.18.2
glob2==0.7
idna==2.9
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
kiwisolver==1.4.7
libarchive-c==2.9
MarkupSafe==1.1.1
matplotlib==3.7.5
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv-full==1.5.3
# Editable install with no version control (mmdet==2.25.1)
-e /home/Adversarial/mmdetection-2.25.1
# Editable install with no version control (mmrotate==0.3.3)
-e /home/Adversarial/mmrotate-0.3.3
mpmath==1.3.0
numpy==1.23.2
nvidia-ml-py==12.570.86
nvitop==1.4.2
olefile==0.46
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pkginfo==1.6.0
platformdirs==4.3.6
prompt-toolkit==3.0.8
psutil==5.7.2
ptyprocess==0.6.0
pycocotools==2.0.7
pycosat==0.6.3
pycparser==2.20
Pygments==2.7.1
pyOpenSSL==19.1.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-etcd==0.4.5
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
ruamel-yaml==0.15.87
scipy==1.10.1
shapely==2.0.7
six==1.14.0
soupsieve==2.0.1
sympy==1.13.3
terminaltables==3.1.10
tomli==2.2.1
torch==1.7.0+cu110
torchaudio==0.7.0
torchelastic==0.2.1
torchgeometry==0.1.2
torchvision==0.8.0
tqdm==4.46.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
yapf==0.40.1
zipp==3.20.2
